+++ resources/views/components/always-responsive-table.blade.php
@@ -1,9 +1,5 @@
-<div class="overflow-x-auto">
-        <div class="min-w-full">
-            <div class="overflow-x-auto p-1">
-                <table {{ $attributes->merge([ 'class' => "min-w-full"]) }}>
-                    {{ $slot }}
-                </table>
-            </div>
-    </div>
+<div class="overflow-x-auto min-w-full">
+        <table {{ $attributes->merge([ 'class' => "min-w-full"]) }}>
+            {{ $slot }}
+        </table>
 </div>